Don't put internal macros, types, variables or functions inside of header Get rid of dependency on BSD type names Remove unused macros Add tests Move initialization of microseconds out of chpl_strptime into DateTime module
---
