--- .vscode/settings.json
@@ -1,8 +1,0 @@
-{
-  "cSpell.words": [
-    "Spinlock",
-    "chpl",
-    "inited",
-    "rtype"
-  ]
-}
